requirements.txt
# Core dependencies
streamlit==1.41.1
pandas==2.2.0
plotly==5.19.0

# Database
cassandra-driver==3.29.0
python-dotenv==1.0.1

# API and HTTP
requests==2.32.3
openai>=1.0.0

# AI/ML
langflow>=0.6.3
langchain>=0.0.335